fixed Gamepad, XboxController
Дa, PYCCЙH

+++ src/interfaces/Gamepad.java
@@ -4,7 +4,7 @@ import Enums.Directions;
 
 public interface Gamepad {
 	
-	public float getAxis(int axis);
+	public double getAxis(int axis);
 	public boolean getButton(int button);
 	public boolean getButtonUp(int button);
 	public boolean getButtonDown(int button);
